# Windows: Faster `getenvW` and a standalone environment variable test