--- a/test/standalone/build.zig.zon
+++ b/test/standalone/build.zig.zon
@@ -96,6 +96,9 @@
         .empty_env = .{
             .path = "empty_env",
         },
+        .env_vars = .{
+            .path = "env_vars",
+        },
         .issue_11595 = .{
             .path = "issue_11595",
         },