@squeek502 2025-03-17
> Oh, one more suggestion: maybe expand the comment on getenvW to say that it does linear searches and if the caller is going to frequently fetch environment variables, it may be worthwhile to use std.process.getEnvMap to be able to query for them more efficiently?

My hunch is that you'd have to be querying them *absurdly* frequently before it starts mattering, and I'm assuming the lack of allocation in `getenvW` makes up for any linear-scan-based-inefficiency. That comment may be appropriate for any of the other environment variable getting functions that take an allocator, but it'd probably still be worth benchmarking to validate.